requirements.txt
fastapi
uvicorn

Jinja2
premailer
minio
pyyaml

# To use Gmail api
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
oauth2client
